testing in progress

+++ kl/domain/fileclient/auth.go
@@ -26,43 +26,49 @@ func (fc *fclient) Logout() error {
 		}
 	}
 	hashConfigPath := path.Join(configPath, "box-hash")
-	if err = os.RemoveAll(hashConfigPath); err != nil {
-		return fn.NewE(err)
+	_, err = os.Stat(hashConfigPath)
+	if err == nil {
+		if err = os.RemoveAll(hashConfigPath); err != nil {
+			return fn.NewE(err)
+		}
 	}
 	vpnConfigPath := path.Join(configPath, "vpn")
-	files, err := os.ReadDir(vpnConfigPath)
-	if err != nil {
-		return fn.NewE(err)
-	}
-	for _, file := range files {
-		_, err := os.Stat(path.Join(vpnConfigPath, file.Name()))
+	_, err = os.Stat(vpnConfigPath)
+	if err == nil {
+		files, err := os.ReadDir(vpnConfigPath)
 		if err != nil {
-			fn.PrintError(err)
-			continue
+			return fn.NewE(err)
 		}
-		content, err := os.ReadFile(path.Join(vpnConfigPath, file.Name()))
-		if err != nil {
-			fn.PrintError(err)
-			continue
-		}
+		for _, file := range files {
+			_, err := os.Stat(path.Join(vpnConfigPath, file.Name()))
+			if err != nil {
+				fn.PrintError(err)
+				continue
+			}
+			content, err := os.ReadFile(path.Join(vpnConfigPath, file.Name()))
+			if err != nil {
+				fn.PrintError(err)
+				continue
+			}
 
-		var data AccountVpnConfig
-		err = json.Unmarshal(content, &data)
-		if err != nil {
-			fn.PrintError(err)
-			continue
-		}
-		data.WGconf = ""
+			var data AccountVpnConfig
+			err = json.Unmarshal(content, &data)
+			if err != nil {
+				fn.PrintError(err)
+				continue
+			}
+			data.WGconf = ""
 
-		modifiedContent, err := json.Marshal(data)
-		if err != nil {
-			fmt.Println(err)
-			continue
-		}
+			modifiedContent, err := json.Marshal(data)
+			if err != nil {
+				fmt.Println(err)
+				continue
+			}
 
-		err = os.WriteFile(path.Join(vpnConfigPath, file.Name()), modifiedContent, 0644)
-		if err != nil {
-			fmt.Println(err)
+			err = os.WriteFile(path.Join(vpnConfigPath, file.Name()), modifiedContent, 0644)
+			if err != nil {
+				fmt.Println(err)
+			}
 		}
 	}
